+++ CHANGELOG.md
@@ -11,6 +11,7 @@ https://github.com/jcorporation/myMPD/
 - Feat: Add option for default behavior on click on tag in browse view #1472
 - Feat: Optionally group songs with empty album tag in a special `Unknown Album` album #1472
 - Feat: Support large images
+- Fix: Do not reset scrolling position on update of lists #1478
 
 ***